refactor UserIndicaterController
