+++ RideTogether/AppDelegate.swift
@@ -27,8 +27,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         
     func application(_ application: UIApplication,
                      didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        // Override point for customization after application launch.
-        
+
         FirebaseApp.configure()
         
         Fabric.with([Crashlytics.self])
@@ -39,7 +38,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         
         IQKeyboardManager.shared().shouldResignOnTouchOutside = true
         
-//        UserDefaults.standard.removeObject(forKey: "UserLogined")
+        // 正式 release 要取消
+        UserDefaults.standard.removeObject(forKey: "UserLogined")
         
         if UserDefaults.standard.value(forKey: "UserLogined") == nil {
